macOS 10.15.1, Xcode 11.2 (iOS, tvOS, watchOS, iOSSupport 13.2)

+++ iOS/PrivateFrameworks/CoreSpeech.framework/CSSPGEndpointAnalyzerDelegate-Protocol.h
@@ -12,6 +12,6 @@
 
 @optional
 - (void)spgEndpointAnalyzerDidDetectEndpoint:(CSSPGEndpointAnalyzer *)arg1;
-- (void)spgEndpointAnalyzer:(CSSPGEndpointAnalyzer *)arg1 hasSilenceScoreEstimate:(float)arg2;
+- (void)spgEndpointAnalyzer:(CSSPGEndpointAnalyzer *)arg1 hasSilenceScoreEstimate:(double)arg2;
 @end
